[BE]feat : 일정 조회 기능 구현 - 미등록 공지 리스트 조회 : 개인 일정 중 미등록된 일정들 - 개인 일정 리스트 조회 : 등록 상태인 개인 일정들 + 상태(해야할일,진행중,완료) 카운트 리스트 - 관리자의 공지 파생 일정 리스트 조회 : 해당 공지사항으로 파생된 일정 리스트 + 상태(등록O, 등록O+완료) 카운트 리스트 - 관리자 개별 부여 일정 리스트 조회 : 개인/팀 으로 관리자가 부여한 일정 리스트 + 상태(등록O, 등록O+완료) 카운트 리스트

+++ ssaffice-BE/src/main/java/com/jetty/ssafficebe/schedule/converter/ScheduleConverter.java
@@ -30,5 +30,5 @@ public interface ScheduleConverter {
     @Mapping(target = "noticeSummaryForList", source = "notice")
     ScheduleSummary toScheduleSummary(Schedule schedule);
 
-    SchedulePageResponse toSchedulePageResponse(Page<ScheduleSummary> scheduleSummaryPage, List<Integer> statusCounts);
+    SchedulePageResponse toSchedulePageResponse(Page<ScheduleSummary> scheduleSummaryPage, List<Long> statusCounts);
 }
